requirements.txt
openai>=1.0.0
anthropic>=0.20.0
google-generativeai>=0.4.0
pandas>=2.0.0
matplotlib>=3.7.0
numpy>=1.24.0
python-dotenv>=1.0.0
requests>=2.31.0
